Split translatable fonts into separate files
If a translation does not use latin letters, there is no need to load
its texture or glyph rects. This way it is also easier to switch the
font for the appropriate locale by editing the fonts.cfg file, as a
workaround before better support for translation fonts is developed.