Fix "box water" bug in Linux with most GL drivers supporting blending ops
The check for GLEW_ARB_imaging doesn't seem to be the most reliable way
to check whether GL_FUNC_ADD and GL_FUNC_REVERSE_SUBTRACT are supported,
as the GL_ARB_imaging extension is missing in at least one system where
these operations are not.

r2608's changes to background.cpp suggest that we can check for GLEW
1.4 and later instead, so replicate those conditionals here.

Note that the fallback path is probably still broken.