Makefile: tidy up the mess involving CFLAGS, CPPFLAGS, and CXXFLAGS
For this purpose, CFLAGS is gone (since there's no use for a C compiler
whatsoever) and its contents moved to the new CXXFLAGS; the previous
contents of CXXFLAGS are now in BASE_CXXFLAGS, which includes all the
stuff that would be expected to be passed "by default". CXXFLAGS may or
may not be used to override it.

Include (-I, -i) options (which are *mandatory* to have and should not
be overridden) are found in INC.

By default, CXXFLAGS and CPPFLAGS are now empty, leaving their contents
to the user.

pkg-config calls have been all collapsed. This allows keeping duplicate
argument to a minimum. However, since some recent versions of SDL_mixer
and SDL_ttf don't install .pc files, we are going to assume their
compiler flags are equivalent to those of SDL_image and SDL and only go
with -lSDL_mixer -lSDL_ttf for the linker.